Change RetentionPolicy to RUNTIME in NotNullByDefault annotation
Motivation:
The NotNullByDefault annotation was originally set with a RetentionPolicy of CLASS, which means it is only available during compile-time. However, we want it to be available during runtime as well, so we can use it for null-checking in Kotlin.

Modifications:
Changed the RetentionPolicy of the NotNullByDefault annotation from CLASS to RUNTIME.

Result:
The NotNullByDefault annotation can now be used for null-checking in Kotlin during both compile-time and runtime.