Fix Github CI when PR is created from Fork
Motivation:
Github actions on PRs created from a fork were failing due to restricted access to repository variables. The problem was that the failing actions were trying to access these variables, which is not allowed.

Modifications:
To fix this issue, I removed the variable access from the Github actions.

Results:
After these changes, the Github actions are now working as expected, even when PRs are created from a fork.

Resolves #70