Deprecate SOCKET_TIMEOUT option
Motivation:
The SOCKET_TIMEOUT option has been identified as a NO-OP, meaning it has
no effect on the underlying connection. To reduce confusion and
potential issues for users, it is necessary to deprecate this option.

Modification:
Added a deprecation annotation to the SOCKET_TIMEOUT option and included
a log warning message to inform users that the option is not supported
and will be ignored.

Result:
The SOCKET_TIMEOUT option is now marked as deprecated, and users will be
notified with a log warning message if they attempt to use it. This
change encourages users to remove any references to SOCKET_TIMEOUT in
their code.