Fixed NPE with `SslMode.TUNNEL` Usage
Motivation:
A NPE was identified when utilizing `SslMode.TUNNEL`. The issue arises when `ConnectionContext#isMariaDb` is invoked from `SslBridgeHandler#isTls13Enabled`, leading to an NPE due to the `ConnectionContext` not being initialized at that time.

Modification:
Do not invoke `ConnectionContext#isMariaDb` when it is not initialized.

Result:
This change addresses the NPE issue, ensuring stability when `SslMode.TUNNEL` is selected.
resolves https://github.com/GoogleCloudPlatform/cloud-sql-jdbc-socket-factory/pull/1828